requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
radon>=6.0.0
pylint>=3.0.0
httpx>=0.27.0
pydantic>=2.0.0
aiosqlite>=0.20.0
langgraph-checkpoint-sqlite>=2.0.0
